create register  test
