+++ app.js
@@ -1,0 +1,29 @@
+require('dotenv').config();
+const express = require('express');
+const cors = require('cors');
+const mongoose = require('mongoose');
+const registerRoute = require('./routes/register');
+
+const loginRoute = require('./routes/login');
+
+const app = express();
+app.use(cors());
+app.use(express.json());
+
+const uri = process.env.ATLAS_URI;
+
+mongoose.connect(uri, { useNewUrlParser: true });
+const connection = mongoose.connection;
+connection.once('open', () => {
+  console.log('MongoDB database connection established succesfully');
+});
+
+app.use(loginRoute);
+app.use(registerRoute);
+
+app.use((err, req, res, next) => {
+  console.log(err);
+  res.status(500).json({ error: err.message });
+});
+
+module.exports = app;
