+++ controllers/user.controller.js
@@ -5,15 +5,14 @@ const Tutor = require('../models/tutor.model');
 const jwt = require('jsonwebtoken');
 
 const getUserData = async (req, res) => {
-    try {
-        const token = req.query.token;
-        console.log(req.query.token)
-        const data = jwt.verify(token, 'secret key');
-        res.json(data);
-    } catch (error) {
-        res.send(500)
-    }
-    
-}
+  try {
+    const token = req.query.token;
+    // console.log(req.query.token)
+    const data = jwt.verify(token, 'secret key');
+    res.json(data);
+  } catch (error) {
+    res.send(500);
+  }
+};
 
-module.exports = { getUserData }
+module.exports = { getUserData };
