+++ server.js
@@ -1,29 +1,3 @@
-require('dotenv').config();
-const express = require('express');
-const cors = require('cors');
-const mongoose = require('mongoose');
-const registerRoute = require('./routes/register');
-
-const loginRoute = require('./routes/login');
-
-const app = express();
-
-app.use(cors());
-app.use(express.json());
-app.use(loginRoute);
-app.use(registerRoute);
-
-const uri = process.env.ATLAS_URI;
-
-mongoose.connect(uri, { useNewUrlParser: true });
-const connection = mongoose.connection;
-connection.once('open', () => {
-  console.log('MongoDB database connection established succesfully');
-});
-
-app.use((err, req, res, next) => {
-  console.log(err);
-  res.status(500).json({ error: err.message });
-});
+const app = require('./app');
 
 app.listen(3001, () => console.log('Server running ...'));
